clean up the api
