+++ bskydata/agents/__init__.py
@@ -1,0 +1,1 @@
+from bskydata.agents.agent import BskyAgent
