+++ bskydata/scrapers/__init__.py
@@ -1,0 +1,3 @@
+from bskydata.scrapers.followers import FollowersScraper
+from bskydata.scrapers.follows import FollowsScraper
+from bskydata.scrapers.profiles import ProfilesScraper
