+++ bskydata/scrapers/profiles.py
@@ -1,83 +1,24 @@
 import time
-import typing as t
 from bskydata.api.client import BskyApiClient
-from atproto import models
-from bskydata.storage.base import DataWriter
+from bskydata.storage.writers.base import DataWriter
 from bskydata.parsers.base import DataParser
 
 
-class ProfileScraper:
-    def __init__(self, bsky_client: BskyApiClient, writer: DataWriter = None, parser: DataParser = None):
+class ProfilesScraper:
+    def __init__(self, 
+                 bsky_client: BskyApiClient, 
+                 writer: DataWriter = None, 
+                 parser: DataParser = None):
         self.bsky_client = bsky_client
         self.writer = writer
         self.parser = parser
     
-    def _fetch_follows(self, actor: str, cursor: t.Union[str, None] = None) -> models.AppBskyGraphGetFollows.Response:
-        params = models.AppBskyGraphGetFollows.Params(actor=actor, limit=100)
-        if cursor:
-            params.cursor = cursor
-        return self.bsky_client.client.app.bsky.graph.get_follows(params)
-
-    def _fetch_followers(self, actor: str, cursor: t.Union[str, None] = None) -> models.AppBskyGraphGetFollowers.Response:
-        params = models.AppBskyGraphGetFollowers.Params(actor=actor, limit=100)
-        if cursor:
-            params.cursor = cursor
-        return self.bsky_client.client.app.bsky.graph.get_followers(params)
-
-    def fetch_all_profiles(self, actors: list, output_file: str = "profiles.json") -> dict:
+    def fetch(self, actors: list, destination: str) -> dict:
         profiles = self.bsky_client.client.get_profiles(actors).model_dump()
         profiles['actors'] = actors
         profiles["created_at"] = time.strftime("%Y-%m-%d %H:%M:%S", time.gmtime())
         if self.parser:
             profiles = self.parser.parse(profiles)
         if self.writer:
-            self.writer.write(profiles, destination=output_file)
+            self.writer.write(profiles, destination=destination)
         return profiles
-
-    def fetch_all_follows(self, actor: str, limit: int = 1000, output_file: str = "follows.json") -> dict:
-        all_follows = []
-        cursor = None
-        limit_check = 0
-        while True:
-            limit_check += 100
-            response = self._fetch_follows(actor, cursor)
-            cursor = response.cursor
-            follows = response.model_dump()['follows']
-            all_follows.extend(follows)
-            if not cursor or limit_check > limit:
-                break
-            time.sleep(2)
-        all_follows_final = {
-            "actor": actor,
-            "created_at":  time.strftime("%Y-%m-%d %H:%M:%S", time.gmtime()),
-            "follows": all_follows
-        }
-        if self.parser:
-            all_follows_final = self.parser.parse(all_follows_final)
-        if self.writer:
-            self.writer.write(all_follows_final, destination=output_file)
-        return all_follows_final
-
-    def fetch_all_followers(self, actor: str, limit: int = 1000, output_file: str = "followers.json") -> dict:
-        all_followers = []
-        cursor = None
-        limit_check = 0
-        while True:
-            limit_check += 100
-            response = self._fetch_followers(actor, cursor)
-            cursor = response.cursor
-            followers = response.model_dump()['followers']
-            all_followers.extend(followers)
-            if not cursor or limit_check > limit:
-                break
-            time.sleep(2)
-        all_followers_final = {
-            "actor": actor,
-            "created_at":  time.strftime("%Y-%m-%d %H:%M:%S", time.gmtime()),
-            "followers": all_followers
-        }
-        if self.parser:
-            all_followers_final = self.parser.parse(all_followers_final)
-        if self.writer:
-            self.writer.write(all_followers_final, destination=output_file)
-        return all_followers_final
